feat(policy): add --before-resolution-policies to the weaver registry check command
