+++ src/registry/check.rs
@@ -2,10 +2,12 @@
 
 //! Check a semantic convention registry.
 
+use std::path::PathBuf;
 use crate::registry::{semconv_registry_path_from, RegistryPath};
 use clap::Args;
 use weaver_cache::Cache;
 use weaver_logger::Logger;
+use weaver_policy_engine::Engine;
 use weaver_resolver::attribute::AttributeCatalog;
 use weaver_resolver::registry::resolve_semconv_registry;
 use weaver_resolver::SchemaResolver;
@@ -25,6 +27,11 @@ pub struct RegistryCheckArgs {
     /// registry is located
     #[arg(short = 'd', long, default_value = "model")]
     pub registry_git_sub_dir: Option<String>,
+
+    /// Optional list of policy files to check against the files of the semantic
+    /// convention registry before the resolution process.
+    #[arg(short = 'b', long)]
+    pub before_resolution_policies: Vec<PathBuf>,
 }
 
 /// Check a semantic convention registry.
@@ -33,6 +40,19 @@ pub(crate) fn command(log: impl Logger + Sync + Clone, cache: &Cache, args: &Reg
     log.loading(&format!("Checking registry `{}`", args.registry));
 
     let registry_id = "default";
+    let policy_engine  = if args.before_resolution_policies.is_empty() {
+        None
+    } else {
+        let mut engine = Engine::new();
+        for policy in &args.before_resolution_policies {
+            engine
+                .add_policy(policy)
+                .unwrap_or_else(|e| {
+                    panic!("Invalid policy file `{:?}`, error: {e}", policy);
+                });
+        }
+        Some(engine)
+    };
 
     // Load the semantic convention registry into a local cache.
     // No parsing errors should be observed.
@@ -41,9 +61,11 @@ pub(crate) fn command(log: impl Logger + Sync + Clone, cache: &Cache, args: &Reg
         semconv_registry_path_from(&args.registry, &args.registry_git_sub_dir),
         cache,
         log.clone(),
+        policy_engine.as_ref(),
     )
     .unwrap_or_else(|e| {
-        panic!("Failed to load and parse the semantic convention registry, error: {e}");
+        e.log(log.clone());
+        std::process::exit(1);
     });
 
     // Resolve the semantic convention registry.
@@ -51,7 +73,8 @@ pub(crate) fn command(log: impl Logger + Sync + Clone, cache: &Cache, args: &Reg
     let registry_path = args.registry.to_string();
     let _ = resolve_semconv_registry(&mut attr_catalog, &registry_path, &semconv_specs)
         .unwrap_or_else(|e| {
-            panic!("Failed to resolve the semantic convention registry.\n{e}");
+            e.log(log.clone());
+            std::process::exit(1);
         });
 
     log.success(&format!(
